finish refactor.
