+++ src/main.rs
@@ -1,35 +1,35 @@
 use std::time::Duration;
 
 use bevy::time::FixedTimestep;
-use plugin::biome::BiomePlugin;
-use system::input::{keyboard_input, scrollwheel_input};
+use crate::plugins::biome::BiomePlugin;
+use crate::systems::input::{keyboard_input, scrollwheel_input};
 use crate::load::{load_font, load_sprites};
-use crate::system::map::generate_map;
-use plugin::monstergenerator::MonsterGeneratorPlugin;
-use system::moverandom::movement_random;
-use system::pause::{on_pause, on_unpause};
-use crate::plugin::main_menu::MainMenusPlugin;
+use crate::systems::map::generate_map;
+use crate::plugins::monstergenerator::MonsterGeneratorPlugin;
+use crate::systems::moverandom::movement_random;
+use crate::systems::pause::{on_pause, on_unpause};
+use crate::plugins::main_menu::MainMenusPlugin;
 
 pub use crate::prelude::*;
-use plugin::startup::StartupPlugin;
-use crate::plugin::button::ButtonPlugin;
-use crate::plugin::click::ClickPlugin;
-use crate::plugin::game_ui::GameUiPlugin;
-use crate::plugin::movetoward::{movement_toward_attackable, MovementPlugin};
-use crate::plugin::needs::{FoodNotifEvent, NeedsPlugin};
-use crate::plugin::seasons::SeasonsPlugin;
-use crate::plugin::selection::SelectionPlugin;
-use crate::plugin::spoilage::SpoilagePlugin;
-use crate::plugin::task::TaskPlugin;
-use crate::plugin::thinking::ThinkingPlugin;
-use crate::system::namegiving::namegiving_system;
-use crate::system::names::names_system;
-use crate::system::statusdisplay::status_display_system;
-use crate::system::text::{text_system, text_test, text_update_system};
-use crate::system::window::{set_window_icon, set_window_maximized};
+use crate::plugins::startup::StartupPlugin;
+use crate::plugins::button::ButtonPlugin;
+use crate::plugins::click::ClickPlugin;
+use crate::plugins::game_ui::GameUiPlugin;
+use crate::plugins::movetoward::{movement_toward_attackable, MovementPlugin};
+use crate::plugins::needs::{FoodNotifEvent, NeedsPlugin};
+use crate::plugins::seasons::SeasonsPlugin;
+use crate::plugins::selection::SelectionPlugin;
+use crate::plugins::spoilage::SpoilagePlugin;
+use crate::plugins::task::TaskPlugin;
+use crate::plugins::thinking::ThinkingPlugin;
+use crate::systems::namegiving::namegiving_system;
+use crate::systems::names::names_system;
+use crate::systems::statusdisplay::status_display_system;
+use crate::systems::text::{text_system, text_test, text_update_system};
+use crate::systems::window::{set_window_icon, set_window_maximized};
 
-mod plugin;
-mod system;
+mod plugins;
+mod systems;
 
 mod prelude;
 mod components;
