+++ src/plugins/click.rs
