+++ src/plugins/monstergenerator.rs
