+++ src/plugins/thinking.rs
