--- src/system/melee_attacker.rs
@@ -1,11 +1,0 @@
-use crate::prelude::*;
-
-
-pub fn melee_attacks(
-    mut commands: Commands,
-    entities: Query<(Entity, &Position, With<MeleeAttacker>)>,
-) {
-    for (entity, position, _) in entities.iter() {
-
-    }
-}
